Factor out recent buffers logic into `RecentBuffersContext`
